Extract common methods from SleepsController to BaseController
